Add AIX support
Like Solaris, AIX can't call syscall.Syscall and need to use sys/unix
functions
