+++ ts_unix.go
@@ -1,4 +1,4 @@
-// +build  freebsd netbsd openbsd
+// +build aix freebsd netbsd openbsd
 
 // Copyright 2014 Oleku Konko All rights reserved.
 // Use of this source code is governed by a MIT
