+++ ts_x.go
@@ -1,4 +1,4 @@
-// +build !windows,!solaris
+// +build !windows,!solaris,!aix
 
 // Copyright 2014 Oleku Konko All rights reserved.
 // Use of this source code is governed by a MIT
